ai_ready_copilot/pyproject.toml
[project]
name = "ai-ready-copilot"
version = "0.1.0"
description = ""
authors = [
    {name = "Julián",email = "julian.garcia@diverger.ai"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "fastapi (>=0.124.0,<0.125.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "ai_ready_copilot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
